Fix cases if wiki page moved multiple times

+++ src/config.py
@@ -81,8 +81,8 @@ class PayloadSetting(BaseSettings):
     enable_wiki_threading: bool = False
     clear_cache_http: bool = False
     clear_cache_wiki: bool = False
-    clear_wiki_changed: int | None = None
-    clear_wiki_moved: int | None = None
+    clear_wiki_changed: float | None = None
+    clear_wiki_moved: float | None = None
     clear_wiki_empty: bool = False
     skip_mapping: bool = False
     skip_quests: bool = False
